AeroLine-Backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.0
psutil==5.9.8
httpx==0.27.0
pyotp==2.9.0
pydantic==2.7.0
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
sqlite3
